+++ crates/lanok/src/lib.rs
@@ -71,7 +71,7 @@ pub use lanok_peer::{Context, Hello, INITIALIZE, INITIALIZED, SimpleServer};
 
 #[cfg(feature = "async")]
 pub use lanok_peer::{
-    AbandonHook, Abandoned, Handler, HandlerFuture, NoHandler, Peer, PeerBuilder, PeerInfo, Router,
+    AbandonHook, Abandoned, Handler, HandlerFuture, NoHandler, Peer, PeerBuilder, Router,
 };
 
 #[cfg(feature = "async")]